errorHandlingSetup

+++ client/src/pages/Profile.jsx
@@ -9,7 +9,7 @@ import {
   setUserState,
   deleteCurrentUser,
 } from "../redux/currentUser/currentUserSlice";
-import format from "date-fns/format";
+
 import moment from "moment";
 
 export const Profile = () => {
@@ -19,34 +19,43 @@ export const Profile = () => {
   const dispatch = useDispatch();
   const navigate = useNavigate();
 
-  const handleLogout = async () => {
-    dispatch(logoutThunk());
-    dispatch(resetNotes());
-    navigate("/login");
-  };
+  // const handleLogout = async () => {
+  //   dispatch(logoutThunk());
+  //   dispatch(resetNotes());
+  //   navigate("/login");
+  // };
 
-  const homeRequest = async () => {
-    const res = await axios.get("/api/users/home");
-    const data = await res.data;
-    try {
-      dispatch(setNotes(JSON.parse(data.session).user.notes));
-      dispatch(setUserState(JSON.parse(data.session).user));
-    } catch (error) {
-      console.log(error);
-      navigate("/login");
-      alert("SESSION EXPIRED");
-    }
-  };
   useEffect(() => {
+    const homeRequest = async () => {
+      const res = await axios.get("/api/users/home");
+      const data = await res.data;
+
+      try {
+        dispatch(setNotes(JSON.parse(data.session).user.notes));
+        dispatch(setUserState(JSON.parse(data.session).user));
+      } catch (error) {
+        console.log(error);
+        navigate("/login");
+        alert("SESSION EXPIRED");
+      }
+    };
+
     homeRequest();
+    // eslint-disable-next-line
   }, []);
 
   const handleDeleteUser = async () => {
-    const res = await dispatch(deleteCurrentUser({ id: currentUser._id }));
-    if (res.type === "deleteUser/fulfilled") {
-      dispatch(logoutThunk());
-      dispatch(resetNotes());
-      navigate("/login");
+    const confrimation = window.confirm(
+      "Are you sure u wanna delete your account? This action is not undoable"
+    );
+
+    if (confrimation) {
+      const res = await dispatch(deleteCurrentUser({ id: currentUser._id }));
+      if (res.type === "deleteUser/fulfilled") {
+        dispatch(logoutThunk());
+        dispatch(resetNotes());
+        navigate("/login");
+      }
     }
   };
